+++ src/Terrabuild.Configuration/Transpiler/Project.fs
@@ -151,7 +151,7 @@ let transpile (blocks: Block list) =
 
             | Extension name ->
                 if builder.Extensions.ContainsKey name then raiseParseError $"duplicated extension '{name}'"
-                let extension = toExtension block
+                let extension = toExtension name block
                 buildProject blocks { builder with Extensions = builder.Extensions |> Map.add name extension }
 
             | Target name ->
@@ -174,4 +174,3 @@ let transpile (blocks: Block list) =
           Targets = Map.empty 
           Locals = Map.empty }
     buildProject blocks builder
-