+++ src/Terrabuild.Configuration/Transpiler/Workspace.fs
@@ -162,7 +162,7 @@ let transpile (blocks: Block list) =
 
             | Extension name ->
                 if builder.Extensions.ContainsKey name then raiseParseError $"duplicated extension '{name}'"
-                let extension = toExtension block
+                let extension = toExtension name block
                 buildWorkspace blocks { builder with Extensions = builder.Extensions |> Map.add name extension }
 
             | UnknownBlock -> raiseParseError $"unexpected block '{block.Resource}'"